TASK: Modify getMidiTickDuration()
Fixed the function so that if a note is played across multiple beats, that the beat the note is played in and the last beat the note is played in are calculated as midi tick values. Then those values are added to full beat lengths of any additional bars in between the two, if any.